==> 03_gxp_genage.Rmd <==
---
title: "Predicting Pro vs Anti Longevity from Gene Expression"
author: "Will Townes"
date: "3/29/2018"
output: html_document
---

Using gene expression alone to predict pro vs anti longevity genes from genage database.

```{r}
#library(doParallel)
#registerDoParallel(makeCluster(detectCores()-1)) #3 cores on mac
library(dplyr)
source("./carets.R") #loads caret, ROCR, ggplot2
spp<-c("yeast","worm")
sp<-spp[1]
features<-c("archs4","gxp")
fts<-features[1]
```

## Data Loading

```{r}
dat<-readRDS(paste0("./data/auto/",sp,".rds"))
#y<-rep("anti",length(dat$y))
#y[dat$y==1]<-"pro"
#y<-factor(y)
y<-dat$y
idx<-dat$cv_folds[[1]]
ytrn<-y[-idx]
ytst<-y[idx]
X<-dat[[fts]]
#mode(X)<-"numeric"
Xtrn<-X[-idx,]
Xtst<-X[idx,]
```

all algorithms

```{r}
mods<-c("kknn","xgbTree","svmRadialSigma","glmnet","nb")
f<-function(m){
  print(m)
  fitcaret(Xtrn,ytrn,m)
}
system.time(res<-sapply(mods,f,simplify=FALSE))
res2<-lapply(res,caret_roc,Xtst,ytst,ytrn)
for(j in names(res2)){
  res2[[j]]$alg<-j
}
pd<-do.call(rbind,res2)
ggplot(subset(pd,fold=="test"),aes(x=fpr,y=tpr,col=alg))+geom_line()+theme_bw()
pd %>% subset(fold=="test") %>% group_by(alg) %>% summarize(auc=roc2auc(fpr,tpr)) %>% arrange(desc(auc))
```

knn classifier

```{r}
fit<-knn(Xtrn,Xtst,ytrn,k=10,prob=TRUE)
probs<-attr(fit,"prob")
probs[fit=="anti"]<-1-probs[fit=="anti"]
preds<-prediction(probs,ytst)
auc.tmp<-performance(preds,"auc")
(auc<-as.numeric(auc.tmp@y.values))

fit<-kknn(y~.,data.frame(y=ytrn,Xtrn),data.frame(y=ytst,Xtst),k=20)
probs<-predict(fit,type="prob")[,2]
performance(prediction(probs,ytst),"auc")

#knn via caret
system.time(fit<-train(Xtrn, ytrn, preProcess=NULL, metric="Kappa", trControl=trainControl("repeatedcv",5,repeats=2),  method="kknn"))
#tuneGrid=expand.grid(kmax=c(5,10,15),distance=c(1,2),kernel="optimal")
print(fit)
perf<-caret_roc(fit,Xtst,ytst,ytrn)
ggplot(perf,aes(x=fpr,y=tpr,col=fold))+geom_line()+theme_bw()+geom_abline(slope=1,intercept=0,alpha=.5)
roc2auc(perf$fpr[perf$fold=="test"],perf$tpr[perf$fold=="test"])
```

xgboost

```{r}
system.time(fit<-train(Xtrn, ytrn, preProcess=NULL, metric="Kappa", trControl=trainControl("repeatedcv",5,repeats=2),  method="xgbTree"))
print(fit)
perf<-caret_roc(fit,Xtst,ytst,ytrn)
ggplot(perf,aes(x=fpr,y=tpr,col=fold))+geom_line()+theme_bw()+geom_abline(slope=1,intercept=0,alpha=.5)
roc2auc(perf$fpr[perf$fold=="test"],perf$tpr[perf$fold=="test"])
```

svm with caret
```{r}
system.time(fit<-train(Xtrn, ytrn, preProcess=NULL, metric="Kappa", trControl=trainControl("repeatedcv",5,repeats=2),  method="svmRadialSigma", prob.model=TRUE))
print(fit)
perf<-caret_roc(fit,Xtst,ytst,ytrn)
ggplot(perf,aes(x=fpr,y=tpr,col=fold))+geom_line()+theme_bw()+geom_abline(slope=1,intercept=0,alpha=.5)
roc2auc(perf$fpr[perf$fold=="test"],perf$tpr[perf$fold=="test"])
```

GLMnet prediction

```{r}
system.time(fit<-train(Xtrn, ytrn, preProcess=NULL, metric="Kappa", trControl=trainControl("repeatedcv",5,repeats=2),  method="glmnet", family="binomial"))
print(fit)
perf<-caret_roc(fit,Xtst,ytst,ytrn)
ggplot(perf,aes(x=fpr,y=tpr,col=fold))+geom_line()+theme_bw()+geom_abline(slope=1,intercept=0,alpha=.5)
roc2auc(perf$fpr[perf$fold=="test"],perf$tpr[perf$fold=="test"])
```

Naive Bayes

```{r}
trc=trainControl("repeatedcv",10,repeats=2)
args<-list(Xtrn,ytrn,preProcess="zv",metric="Kappa",trControl=trc,method="nb")
args$preProcess<-"nzv"
tg=expand.grid(fL=c(0,0.5,1.0),usekernel=TRUE,adjust=c(0,0.5,1.0))
args$tuneGrid<-tg
args$prior<-table(ytrn)/length(ytrn)
res<-do.call(train,args)

#manually
library(klaR)
bad<-nearZeroVar(X)
Xtrn2<-Xtrn[,-bad]
Xtst2<-Xtst[,-bad]
res<-NaiveBayes(Xtrn2,ytrn,usekernel=TRUE)
preds<-predict(res,Xtst2) #they are all NaN

library(naivebayes) #devtools::install_github("willtownes/naivebayes")
res<-naive_bayes(Xtrn,ytrn,usekernel=TRUE)
preds<-predict(res,Xtst,type="prob")
res<-train(Xtrn,ytrn,preProcess="nzv",metric="Kappa",trControl=trainControl("repeatedcv",3,repeats=1),method="naive_bayes",tuneGrid=expand.grid(laplace=0,usekernel=TRUE,adjust=c(0.5,1)))
preds<-predict(res,Xtst,type="prob")

print(res)
perf<-caret_roc(res,Xtst,ytst,ytrn)
ggplot(perf,aes(x=fpr,y=tpr,col=fold))+geom_line()+theme_bw()+geom_abline(slope=1,intercept=0,alpha=.5)
roc2auc(perf$fpr[perf$fold=="test"],perf$tpr[perf$fold=="test"])
```
