==> 05_go_gxp_genage.Rmd <==
---
title: "Predicting Survival from GO terms and gene expression"
author: "Will Townes"
date: "6/29/2018"
output: html_document
---

```{r}
library(SummarizedExperiment)
library(class)
library(glmnet)
library(caret)
library(ROCR)
source("util.R") #this loads dplyr too
spp<-c("yeast","worm")
sp<-spp[1]
if(sp=="yeast"){
  mutants<-"responsive"
  measured_genes<-TRUE
}
archs4<-TRUE
```

## Data Loading

Merge GO terms, gene expression, and survival Data

```{r}
normalize_func<-function(X){
  #X either a counts matrix or counts scaled by gene length
  log2(1+t(1e6*t(X)/colSums(X)))
}
sv<-load_genage(sp)
rownames(sv)<-toupper(sv$symbol)
Xgo<-load_goterms(sp)
if(archs4){
  se<-readRDS(paste0("./data/archs4",sp,".rds"))
  X<-normalize_func(assay(se,"counts")/rowData(se)$length)
} else {
  if(sp=="yeast"){
    if(mutants=="all"){
      ge<-readRDS("data/deleteome_all_mutants_controls.rds")
    } else if(mutants=="responsive"){
      ge<-readRDS("data/deleteome_responsive_mutants.rds")
    } else { stop("invalid mutant") }
    if(measured_genes==TRUE){
      X<-assay(ge,"logratio")#[sv$symbol,] #sort order match
    } else {
      rownames(ge)<-paste0("expr_",rownames(ge))
      colnames(ge)<-colData(ge)$pert_geneSymbol
      X<-t(assay(ge,"logratio"))#[,sv$symbol])
    }
  } else if(sp=="worm") {
    X<-read.table("./data/worm_scrna_atlas_merged_umi.txt")
    X<-normalize_func(X)
  }
}
gg<-intersect(rownames(Xgo),intersect(rownames(sv),rownames(X)))
X<-cbind(as.matrix(X[gg,]),Xgo[gg,])
X<-X[,apply(X,2,sd)>0] #remove cells with no variation
#remove samples with fewer than 10 nonzero genes
X<-X[,colSums(X>0)>=10]
dim(X)
sv<-sv[gg,]
y<-sv$pro_longevity
names(y)<-rownames(X)
# train/test split
train_idx<-drop(createDataPartition(y,p=.75,list=FALSE))
ytrn<-y[train_idx]
ytst<-y[-train_idx]
```

knn classifier

```{r}
Xtrn<-X[train_idx,]
Xtst<-X[-train_idx,]
fit<-knn(Xtrn,Xtst,ytrn,k=10,prob=TRUE)
probs<-attr(fit,"prob")
probs[fit==0]<-1-probs[fit==0]
preds<-prediction(probs,ytst)
auc.tmp<-performance(preds,"auc")
(auc<-as.numeric(auc.tmp@y.values)) #0.6
```

regression-type models

```{r}
pca<-FALSE
if(pca){
  Xs<-X %*% svd(X,nu=500,nv=500)$v #if PCA adjustment
  colnames(Xs)<-paste0("pca_",seq.int(500))
} else {
  #random subset of columns or use PCA to reduce dims
  nsample<-min(2500,ncol(X))
  Xs<-X[,sample.int(ncol(X),nsample)]
}
Xtrn<-Xs[train_idx,]
Xtst<-Xs[-train_idx,]
```

fixed prediction probability= probability of positive class

```{r}
probs<-rep(mean(ytrn==1),length(ytst))
auc.tmp<-performance(prediction(probs,ytst),"auc")
(auc<-as.numeric(auc.tmp@y.values))
```

fixed prediction probability= all major (negative) class

```{r}
probs<-rep(0,length(ytst))
auc.tmp<-performance(prediction(probs,ytst),"auc")
(auc<-as.numeric(auc.tmp@y.values)) 
```

gradient boosted machines

```{r}
system.time(res2<-train(Xtrn,factor(ytrn), method="gbm", distribution="bernoulli", preProcess=NULL, verbose=FALSE, metric="Kappa", trControl=trainControl("cv",5), tuneGrid=expand.grid(interaction.depth=c(1,2),n.trees=c(800,1600),shrinkage=.005,n.minobsinnode=10)))
print(res2)
probs<-predict(res2,newdata=Xtst,type="prob")[,"1"]
preds<-prediction(probs,ytst)
auc.tmp<-performance(preds,"auc")
(auc<-as.numeric(auc.tmp@y.values))
```

svm with caret
```{r}
system.time(res3<-train(Xtrn,factor(ytrn), method="svmRadial", distribution="bernoulli", preProcess=NULL, verbose=FALSE, metric="Kappa", trControl=trainControl("cv",5),prob.model=TRUE))
print(res3)
probs<-predict(res3,newdata=Xtst,type="prob")[,"1"]
preds<-prediction(probs,ytst)
auc.tmp<-performance(preds,"auc")
(auc<-as.numeric(auc.tmp@y.values))
plot(performance(preds, measure = "tpr", x.measure = "fpr"),main="SVM")
abline(v=1-.803,lty=2)
abline(h=0.6,lty=2)
```

GLMnet prediction

```{r}
#pp<-preProcess(Xtrn,method=c("center","scale"))
#Xtrn2<-cbind(1,predict(pp,Xtrn))
#Xtst2<-cbind(1,predict(pp,Xtst))
Xtrn2<-cbind(1,Xtrn)
Xtst2<-cbind(1,Xtst)

#folds<-createFolds(ytrn)
#res<-train(mean_lifespan~.,data=d[train_idx,],preProcess=c("center","scale"))

alpha<-if(pca){ 0.0 }else{ .9 }
res<-cv.glmnet(Xtrn2,ytrn,family="binomial",alpha=alpha)
probs<-predict(res, newx=Xtst2, s="lambda.min", type="response")
#saveRDS(res,file="glmnet_fit.rds")
plot(res)
preds<-prediction(probs,ytst)
auc.tmp<-performance(preds,"auc")
(auc<-as.numeric(auc.tmp@y.values))
```
