==> 07_setup_autorun.R <==
#>>time Rscript 07_setup_autorun.R
#This script should be run after 01_data_loading.Rmd
#library(parallel)
source("./carets.R")

fp <- file.path

mkdir_p <- function(d) {
  if (!dir.exists(d)) {
    dir.create(d, recursive = TRUE)
  }
}

make_predictors <- function(features, dat) {
  #features a string like "go" or "go_archs4"
  #dat a list generated by load_data function from 01_data_loading.Rmd
  #output: design matrix X suitable for regression modeling
  fts <- unlist(strsplit(features, "_"))
  do.call(cbind, dat[fts])
}

caret_all <- function(Xtrn, ytrn) {
  algs <- c("naive_bayes", "kknn", "xgbTree", "svmRadialSigma", "glmnet")
  f <- function(a, Xtrn, ytrn) {
    fitcaret(Xtrn, ytrn, a)
  }
  fits <- sapply(algs, f, Xtrn, ytrn, simplify = FALSE)
}

dat2roc <- function(
  features,
  fold,
  dat,
  basedir,
  usecache = TRUE,
  verbose = TRUE
) {
  #features is a string saying what predictors to use (eg go_gxp)
  #fold is which CV fold to use from dat (an integer in 1...5)
  #dat is the list of all data
  #caches the list of fitted models (all algs) under <basedir>/<features>/carets.rds
  #writes the tpr and fpr (ROC) to <basedir>/<features>/roc.txt
  #basedir is usually something like "./results/yeast"
  pth <- fp(basedir, features) #eg "./results/yeast/go_gxp"
  if (verbose) {
    print(paste(pth, "fold:", fold))
  }
  mkdir_p(pth)
  cachepath <- fp(pth, paste0("carets", fold, ".rds"))
  test_idx <- dat$cv_folds[[fold]]
  ytrn <- dat$y[-test_idx]
  ytst <- dat$y[test_idx]
  X <- make_predictors(features, dat)
  mode(X) <- "numeric"
  Xtrn <- X[-test_idx, ]
  Xtst <- X[test_idx, ]
  if (usecache && file.exists(cachepath)) {
    #if(verbose){ print("reusing cached caret fits") }
    fits <- readRDS(cachepath)
  } else {
    #if(verbose){ print("fitting caret models") }
    fits <- caret_all(Xtrn, ytrn)
    saveRDS(fits, file = cachepath)
  }
  rocfunc <- function(a) {
    d <- caret_roc(fits[[a]], Xtst, ytst, ytrn)
    d$alg <- a
    d
  }
  res <- do.call(rbind, lapply(names(fits), rocfunc))
  res$cv_fold <- fold
  write.table(
    res,
    file = fp(pth, paste0("roc", fold, ".txt")),
    quote = FALSE,
    row.names = FALSE
  )
}

master <- function(sp, ft, fold) {
  #sp="worm" or "yeast"
  #ft= "go", "gxp", "go_gxp" etc
  #runs all algorithms for the species/features combo and saves to file
  sp <- as.character(sp)
  ft <- as.character(ft)
  dat <- readRDS(fp("./data/auto", paste0(sp, ".rds")))
  dat2roc(ft, fold, dat, fp("./results", sp), usecache = TRUE, verbose = TRUE)
}

species <- c("worm", "yeast")
for (sp in species) {
  mkdir_p(fp("./results", sp))
}
fts <- c("go", "archs4", "gxp", "go_archs4", "go_gxp")
cv_folds <- 1:5
#yeast gxp is deleteome measured genes
#worm gxp is merged UMIs of worm cell atlas
atlas <- expand.grid(fold = cv_folds, fts = fts, species = species)
parallel::mcmapply(master, atlas$species, atlas$fts, atlas$fold, mc.cores = 4)
#mapply(master,atlas$species,atlas$fts)
