==> 08_auc_graphs.Rmd <==
---
title: "Make Graphs Summarizing AUC Patterns"
author: "Will Townes"
date: "10/11/2018"
output: html_document
---

```{r}
library(tidyverse)
library(pheatmap)
theme_set(theme_bw())
source("./carets.R")
sp<-TRUE #save plots?
fp<-file.path
bp<-"./results"
pth<-fp(bp,"plots")
if(!dir.exists(pth)){
  dir.create(pth,recursive=TRUE)
}
ggs<-function(plt,w=6,h=4,...){
  if(sp){ ggsave(file=fp(pth,plt),width=w,height=h,...) }
}
```

Get list of hyperparameters for each algorithm

```{r}
x<-readRDS(fp(bp,"yeast/go/carets1.rds"))
res<-lapply(x,function(t){t$results[,1:(ncol(t$results)-4)]})
y<-readRDS(fp(bp,"worm/archs4/carets2.rds"))
res2<-lapply(y,function(t){t$results[,1:(ncol(t$results)-4)]})
res_tree<-lapply(res$xgbTree,unique)
res2_tree<-lapply(res$xgbTree,unique)
```

```{r}
#load all ROC results
species<-c("worm","yeast")
fts<-c("go","archs4","gxp","go_archs4","go_gxp")
atlas<-expand.grid(features=fts,species=species,cv_folds=1:5,stringsAsFactors=FALSE)
load_data<-function(sp,ft,fld){
  pth<-file.path(bp,sp,ft,paste0("roc",fld,".txt"))
  d<-read.table(pth,header=TRUE)
  d$species<-sp
  d$features<-ft
  d$cv_fold<-fld
  d
}
pd<-do.call(rbind,mapply(load_data,atlas$species,atlas$features,atlas$cv_folds,SIMPLIFY=FALSE))
rownames(pd)<-NULL
auc0<-pd %>% group_by(species,features,alg,fold,cv_fold) %>% summarize(auc=roc2auc(fpr,tpr))
auc0$alg<-plyr::mapvalues(auc0$alg,from=c("glmnet","svmRadialSigma","xgbTree","naive_bayes","kknn"),to=c("pglm","svm","xgb","nb","knn"))
colnames(auc0)[colnames(auc0)=="alg"]<-"algorithms"
auc<-subset(auc0,fold=="test")
auc$features<-factor(auc$features,levels=c("archs4","gxp","go","go_archs4","go_gxp"),ordered=TRUE)
auc$algorithms<-factor(auc$algorithms)
```

pairwise comparison of all algorithms. Focusing only on test set AUC, for each combination of species, features set, and cross validation fold, rank the algorithms according to their AUC. Then, for each pairwise comparison of algorithms, compute the fraction of folds where one algorithm was superior to the other. Order the algorithms by the minimum of these proportions. The idea is that a consistently good algorithm is better than all other algorithms uniformly.

```{r}
alg_ranking<-function(auc){
  #algs is a data frame with columns "algorithms" and "auc"
  #creates a table with pairwise comparison of which algorithms
  #had better performance across experimental conditions
  algs<-levels(auc$algorithms)
  pc<-pivot_wider(auc,names_from="algorithms",values_from="auc")
  rk<-t(apply(pc[,algs],1,rank))
  pc2<-matrix(0,nrow=length(algs),ncol=length(algs))
  rownames(pc2)<-colnames(pc2)<-algs
  for(i in seq_along(algs)){
    for(j in seq_along(algs)){
      #fraction of folds where 'i' was better than 'j'
      pc2[i,j]<-mean(rk[,i]>rk[,j])
    }
  }
  o<-order(apply(pc2,1,function(t){min(t[t>0])}),decreasing=TRUE)
  pc2[o,o]
}

pc2<-alg_ranking(auc)
write.table(pc2,file="./results/alg_table.txt",quote=FALSE)

pc2w<-alg_ranking(subset(auc,species=="worm"))  
write.table(pc2w,file="./results/alg_table_worm.txt",quote=FALSE)

pc2y<-alg_ranking(subset(auc,species=="yeast"))  
write.table(pc2y,file="./results/alg_table_yeast.txt",quote=FALSE)

print(pc2); print(pc2w); print(pc2y)

#order levels of algs factor based on performance
auc$algorithms<-factor(auc$algorithms,levels=rev(rownames(pc2)),ordered=TRUE)
```

turn pairwise table into a plot

```{r}
pc2<-as.matrix(read.table("./results/alg_table.txt",header=TRUE))
pheatmap(pc2,display_numbers = TRUE,cluster_rows=FALSE,cluster_cols=FALSE,color = colorRampPalette(c('white','orange'))(100),fontsize = 20,angle_col=0,legend=FALSE,filename=fp(pth,"alg_table_heatmap.pdf"))
#EPS save for manuscript, have to do manual since pheatmap doesn't support
#setEPS()
#postscript(fp(pth,"fig1.eps"),width=5,height=4.5)
pheatmap(pc2,display_numbers = TRUE,cluster_rows=FALSE,cluster_cols=FALSE,color = colorRampPalette(c('white','orange'))(100),fontsize = 20,angle_col=0,legend=FALSE)
#dev.off()
```

what is the best combination of features and algorithm within each species?

```{r}
auc2<- auc %>% group_by(species,features,algorithms) %>% summarize(auc_lo=min(auc),auc_avg=mean(auc),auc_hi=max(auc))

auc_worm<-subset(auc2,species=="worm")
best_worm<-which.max(auc_worm$auc_avg)
auc_worm[best_worm,]

auc_yeast<-subset(auc2,species=="yeast")
best_yeast<-which.max(auc_yeast$auc_avg)
auc_yeast[best_yeast,]
```

Draw ROC for each species for best performing model (pglm with GO+ARCHS4 for worm, pglm with GO+GXP for yeast).

```{r}
#load full results not just presummarized AUC values
f<-function(k,pth){
  roc<-read.table(fp(pth,paste0("roc",k,".txt")),header=TRUE)
  subset(roc,alg=="glmnet" & fold=="test")
}
roc_w<-do.call(rbind,lapply(as.list(1:5),f,pth=fp(bp,"worm/go_archs4")))
roc_w$species<-"worm"
roc_y<-do.call(rbind,lapply(as.list(1:5),f,pth=fp(bp,"yeast/go_gxp")))
roc_y$species<-"yeast"
roc<-rbind(roc_w,roc_y)
ggplot(roc,aes(x=fpr,y=tpr,group=cv_fold))+geom_segment(x=0,y=0,xend=1,yend=1,linetype="dotted",colour="gray",size=.5)+geom_line()+xlab("false positive rate")+ylab("true positive rate")+facet_wrap(~species)
ggs("roc_best.pdf")
ggs("fig3.eps",w=5.2,h=3.5,dpi=300,device="eps")
```

overall summary of all results

```{r}
cbb<-c("#E69F00","#56B4E9","#009E73","#0072B2","#D55E00","#CC79A7","#F0E442","#000000")
ggplot(auc,aes(x=algorithms,y=auc,fill=features,colour=features))+geom_boxplot()+facet_wrap(~species)+scale_fill_manual(values=cbb)+scale_colour_manual(values=cbb)+ylab("area under ROC")#+theme(axis.text.x=element_text(angle = 30, hjust = 1))
ggs("auc_all_xalg.pdf")
ggs("figS1.eps",w=5.2,h=3.5,dpi=300,device="eps")

ggplot(auc,aes(x=features,y=auc,fill=algorithms,colour=algorithms))+geom_boxplot()+facet_wrap(~species)+scale_fill_manual(values=cbb)+scale_colour_manual(values=cbb)+ylab("area under ROC")+theme(axis.text.x=element_text(angle = 30, hjust = 1))
ggs("auc_all_xfeat.pdf")
ggs("figS2.eps",w=5.2,h=3.5,dpi=300,device="eps")

#subset on best performing algorithms only
auc2<-subset(auc,algorithms %in% c("svm","pglm"))
auc2$algorithms<-factor(auc2$algorithms,levels=c("pglm","svm"))
ggplot(auc2,aes(x=features,y=auc,fill=algorithms,colour=algorithms))+geom_boxplot()+facet_wrap(~species)+scale_fill_manual(values=cbb[5:4])+scale_colour_manual(values=cbb[5:4])+ylab("area under ROC")+theme(axis.text.x=element_text(angle = 30, hjust = 1),legend.position="top")
ggs("auc_svm_pglm.pdf")
ggs("fig2.eps",w=5.2,h=3.5,dpi=300,device="eps")
```
