==> 09_predict_mccormick.Rmd <==
---
title: "GO GXP McCormick"
author: "Will Townes"
date: "11/14/2018"
output: html_document
---

Comparing model prediction of pro-longevity vs observed change in RLS for McCormick data. Also (below), interpreting the coefficients of GLMNET fit to GO-only.

```{r}
library(SummarizedExperiment)
library(glmnet) 
library(ggplot2)
library(GO.db)
#library(biomaRt)
source("./util.R") #loads dplyr and Matrix
fp<-file.path
bp<-"./results/auto"
pp<-fp(bp,"plots")
if(!dir.exists(pp)){
  dir.create(pp,recursive=TRUE)
}
tp<-fp(bp,"tables")
if(!dir.exists(tp)){
  dir.create(tp,recursive=TRUE)
}
```

```{r}
species<-c("worm","yeast")

normalize_gxp_counts<-function(X){
  #X either a counts matrix or counts scaled by gene length
  log2(1+t(1e6*t(X)/colSums(X)))
}

preprocess<-function(X,binary=FALSE,min_nonzero=1){
  X<-as.matrix(X)
  X<-X[,colSums(X!=0)>=min_nonzero]
  sds<-apply(X,2,sd)
  X<-X[,sds>1e-12]
  if(!binary){ 
    X<-.5*scale(X) 
  }
  X
}

load_data<-function(sp){
  sv<-load_genage(sp)
  rownames(sv)<-toupper(sv$symbol)
  archs4<-readRDS(paste0("./data/archs4",sp,".rds"))
  archs4<-assay(archs4,"counts")/rowData(archs4)$length
  archs4<-normalize_gxp_counts(archs4)
  go<-load_goterms(sp)
  gg<-intersect(rownames(go),rownames(archs4))
  id_trn<-intersect(gg,rownames(sv))
  y0<-sv[id_trn,"pro_longevity"]
  y<-rep("anti",length(y0))
  y[y0==1]<-"pro"
  y<-factor(y)
  names(y)<-id_trn
  archs4<-preprocess(archs4[gg,])
  go<-preprocess(go[gg,],binary=TRUE)
  X<-cbind(archs4,go)
  id_tst<-setdiff(rownames(X),id_trn)
  list(y=y,Xtrn=X[id_trn,],Xtst=X[id_tst,])#,cv_folds=cv_folds)
}

get_gene_info<-function(g,sp=c("yeast","worm")){
  sp_match<-list(yeast=c("scerevisiae","ensembl_gene_id"),
                  worm=c("celegans","wormbase_gene"))
  x<-biomaRt::useMart("ensembl", host="http://jul2018.archive.ensembl.org", dataset=paste0(sp_match[[sp]][1],"_gene_ensembl"))
  #gname<-sp_match[[sp]][2] #eg, "wormbase_gene"
  z<-biomaRt::getBM(c("external_gene_name","external_gene_source","ensembl_gene_id","gene_biotype","description"), filters="external_gene_name",values=g,mart=x)
}
```

Apply glmnet to combined GO and ARCHS4 dataset, cross validate 21 different alpha values from 0 (ridge) to 1 (lasso). glmnet automatically searches grid of 97 different lambda values

```{r}
#sp<-"worm"
for(sp in species){
  print(sp)
  dat<-load_data(sp)
  cv_folds<-caret::createFolds(dat$y,k=5)
  for(i in seq_along(cv_folds)){
    cv_folds[[i]]<-data.frame(id=cv_folds[[i]],fold=i)
  }
  cv_folds<-do.call(rbind,cv_folds)
  o<-order(cv_folds$id)
  cv_folds<-cv_folds[o,]
  foldid<-cv_folds$fold
  alphas<-seq(from=0,to=1,by=.05)
  f<-function(a){
    cv.glmnet(dat$Xtrn,dat$y,alpha=a,foldid=foldid,family="binomial",standardize=FALSE)
  }
  system.time(res<-lapply(alphas,f))
  best<-list(alpha=NA,cvm=Inf)
  for(i in seq_along(alphas)){
    a<-alphas[i]
    min_cvm<-min(res[[i]]$cvm)
    if(min_cvm<best$cvm){
      best$alpha<-a; best$cvm<-min_cvm
    }
  }
  print(best)
  fit<-res[[which(alphas==best$alpha)]]
  g<-function(fit,type=c("train","test")){
    type<-match.arg(type)
    if(type=="train"){
      preds<-predict(fit,dat$Xtrn,type="response",s="lambda.min")
    } else {
      preds<-predict(fit,dat$Xtst,type="response",s="lambda.min")
    }
    res<-data.frame(gene=rownames(preds),prob_pro_longevity=preds[,1])
    res$genage_annotated<-(type=="train")
    res
  }
  preds<-rbind(g(fit,"train"),g(fit,"test"))
  ofile<-fp(bp,paste0("final_preds_",sp,".txt"))
  write.table(preds,file=ofile,quote=FALSE,row.names=FALSE)
}
```

For worm, the optimal alpha was .05 (mostly ridge regression). For yeast, the optimal alpha was 0.5 (even mix between ridge and lasso).

```{r}
worm<-read.table("./results/auto/final_preds_worm.txt",header=TRUE)
worm$species<-"worm"
yeast<-read.table("./results/auto/final_preds_yeast.txt",header=TRUE)
yeast$species<-"yeast"
pd<-rbind(worm,yeast)
pd$ga<-"non-genage"
pd$ga[pd$genage_annotated]<-"genage"
ggplot(pd,aes(x=prob_pro_longevity))+geom_histogram(color="black",fill="white")+facet_grid(ga~species,scales="free_y")+theme_bw()+xlab("predicted probability gene is pro-longevity")+expand_limits(x=c(0,1))
ggsave(fp(pp,"final_preds_hist.pdf"),width=5,height=5)
```

Compare pro-longevity probability with replicative lifespan from McCormick (yeast only)

```{r}
sv<-read.csv("data/mccormick_all.csv",stringsAsFactors=FALSE)
sv<-subset(sv,numgenes==1)
pd2<-merge(subset(pd,species=="yeast"),sv,by.x="gene",by.y="set_genotype")

(plt<-ggplot(pd2,aes(x=prob_pro_longevity,y=lifespan_change))+geom_point(size=.5,alpha=.9)+theme_bw() +xlab("predicted probability gene is pro-longevity") +ylab("change in replicative lifespan after knockout") +geom_smooth(color="blue",size=.5,se=FALSE)+facet_wrap(~ga))
#(plt2<-ggExtra::ggMarginal(plt,type="histogram",fill="white",bins=100))
ggsave(fp(pp,"probs_vs_rls.pdf"),width=7,height=4)

#same plot but focus only on top 20 and bottom 20 genes
f<-function(d,G=20){
  o<-order(d$prob_pro_longevity)
  d<-d[o,]
  d$effect<-"pro-longevity"
  d$effect[1:G]<-"anti-longevity"
  rbind(head(d,G),tail(d,G))
}
pd3<-rbind(f(subset(pd2,ga=="genage")),f(subset(pd2,ga=="non-genage")))
colnames(pd3)[colnames(pd3)=="prob_pro_longevity"]<-"prob"
ggplot(pd3,aes(x=effect,y=lifespan_change,colour=prob))+geom_jitter(width=.2,size=3)+facet_wrap(~ga)+theme_bw()+scale_colour_continuous(low="red",high="blue")+xlab("predicted effect")+ylab("change in replicative lifespan after knockout")
ggsave(fp(pp,"probs_vs_rls_G20.pdf"),width=7,height=4)

#t-test for genage and non-genage
t.test(lifespan_change~effect,subset(pd3,ga=="genage"),alternative="greater")#,var.equal=TRUE)
t.test(lifespan_change~effect,subset(pd3,ga=="non-genage"),alternative="greater")#,var.equal=TRUE)
```

I did one-sided t-tests since it has more power and we would never expect knocking out the pro-longevity genes to increase lifespan. If you like I can also do the two-sided tests. I did not assume the two groups have equal variance. The conclusion is, knocking out predicted anti-longevity genes in yeast significantly increases the replicative lifespan change (relative to control) compared to knocking out predicted pro-longevity genes. The level of confidence is much higher for genage genes compared to non-genage genes.

```
# genage genes
	Welch Two Sample t-test

data:  lifespan_change by effect
t = 5.0313, df = 35.094, p-value = 7.251e-06
alternative hypothesis: true difference in means is greater than 0
95 percent confidence interval:
 6.869778      Inf
sample estimates:
mean in group anti-longevity  mean in group pro-longevity 
                    3.824407                    -6.518328 
# non-genage genes                    
	Welch Two Sample t-test

data:  lifespan_change by effect
t = 2.0046, df = 36.721, p-value = 0.02621
alternative hypothesis: true difference in means is greater than 0
95 percent confidence interval:
 0.4100382       Inf
sample estimates:
mean in group anti-longevity  mean in group pro-longevity 
                   0.7187705                   -1.8726971               
```


extract out the top ten pro and anti-longevity genes that were not in genage

```{r}
d<-subset(pd,!genage_annotated)
#sp<-"worm"
for(sp in species){
  ds<-subset(d,species==sp)
  pro<-head(ds[order(ds$prob_pro_longevity,decreasing=TRUE),],10)
  pro<-pro[,c("gene","prob_pro_longevity")]
  colnames(pro)[2]<-"predicted_probability"
  pro$aging_effect<-"pro_longevity"
  pro$predicted_probability<-round(pro$predicted_probability,3)
  anti<-head(ds[order(ds$prob_pro_longevity),],10)
  anti<-anti[,c("gene","prob_pro_longevity")]
  colnames(anti)[2]<-"predicted_probability"
  anti$aging_effect<-"anti_longevity"
  anti$predicted_probability<-round(1-anti$predicted_probability,3)
  res<-rbind(pro,anti)
  ginfo<-get_gene_info(as.character(res$gene),sp=sp)
  colnames(ginfo)[colnames(ginfo)=="external_gene_name"]<-"gene"
  ginfo$gene<-toupper(ginfo$gene)
  res<-merge(res,ginfo,by="gene",sort=FALSE)
  ofile<-fp(tp,paste0("topgenes_",sp,".csv"))
  #write.table(res,file=ofile,quote=FALSE,row.names=FALSE)
  write.csv(res,file=ofile,row.names=FALSE)
}
```

reformat tables for manuscript

```{r}
for(sp in c("worm","yeast")){
  d<-read.csv(fp(tp,paste0("topgenes_",sp,".csv")))
  for(ae in c("pro","anti")){
    d2<-subset(d,aging_effect==paste0(ae,"_longevity"))
    d2<-d2[,c("gene","predicted_probability","ensembl_gene_id","description")]
    colnames(d2)<-c("Gene","Prob","ID","Description")
    if(sp=="yeast"){
      z<-strsplit(as.character(d2$Description),"; ",fixed=TRUE)
      d2$Description<-vapply(z,function(x){ifelse(length(x)>0,x[1],"")},FUN.VALUE="")
    }
    d2$Description<-gsub("\\s+?\\[Source:.*","",d2$Description,perl=TRUE)
    d2$Description<-gsub("Major sperm protein.*","Major sperm protein",d2$Description,perl=TRUE)
    ofile<-fp(tp,paste0("topgenes_",sp,"_",ae,"_longevity.csv"))
    write.csv(d2,file=ofile,row.names=FALSE,quote=FALSE)
  }
}
```

Apply glmnet to GO dataset, cross validate 21 different alpha values from 0 (ridge) to 1 (lasso). glmnet automatically searches grid of 97 different lambda values

```{r}
load_data_go_only<-function(sp){
  sv<-load_genage(sp)
  rownames(sv)<-toupper(sv$symbol)
  go<-load_goterms(sp)
  id_trn<-intersect(rownames(go),rownames(sv))
  y0<-sv[id_trn,"pro_longevity"]
  y<-rep("anti",length(y0))
  y[y0==1]<-"pro"
  y<-factor(y)
  names(y)<-id_trn
  X<-preprocess(go,binary=TRUE)
  id_tst<-setdiff(rownames(X),id_trn)
  list(y=y,Xtrn=X[id_trn,],Xtst=X[id_tst,])#,cv_folds=cv_folds)
}

#sp<-"worm"
for(sp in species){
  print(sp)
  dat<-load_data_go_only(sp)
  cv_folds<-caret::createFolds(dat$y,k=5)
  for(i in seq_along(cv_folds)){
    cv_folds[[i]]<-data.frame(id=cv_folds[[i]],fold=i)
  }
  cv_folds<-do.call(rbind,cv_folds)
  o<-order(cv_folds$id)
  cv_folds<-cv_folds[o,]
  foldid<-cv_folds$fold
  alphas<-seq(from=0,to=1,by=.05)
  f<-function(a){
    cv.glmnet(dat$Xtrn,dat$y,alpha=a,foldid=foldid,family="binomial",standardize=FALSE)
  }
  toc<-system.time(res<-lapply(alphas,f))
  print(toc)
  best<-list(alpha=NA,cvm=Inf)
  for(i in seq_along(alphas)){
    a<-alphas[i]
    min_cvm<-min(res[[i]]$cvm)
    if(min_cvm<best$cvm){
      best$alpha<-a; best$cvm<-min_cvm
    }
  }
  print(best)
  fit<-res[[which(alphas==best$alpha)]]
  ofile<-fp(bp,paste0("go_only_glmnet_fit_",sp,".rds"))
  saveRDS(fit,file=ofile)
}
```

For GO predictors only, the best CV was achieved with alpha=0.15 for worm and alpha=0.1 for yeast. We now extract the most informative predictors by thresholding the coefficients and reporting on the odds ratio scale

```{r}
spp<-c("worm","yeast")
for(sp in spp){
  fit<-readRDS(fp(bp,paste0("go_only_glmnet_fit_",sp,".rds")))
  b<-coef(fit,s="lambda.min")[-1,1]
  o<-order(abs(b),decreasing=TRUE)
  res<-head(b[o],20)
  res2<-AnnotationDbi::select(GO.db,keys=names(res),columns=c("ONTOLOGY","TERM"),keytype="GOID")
  res<-cbind(res,res2)
  res<-res[,c(2,1,3,4)]
  colnames(res)<-c("ID","Effect","Type","Description")
  res$Description<-gsub(",",";",res$Description,fixed=TRUE)
  res$Effect<-signif(res$Effect,2)
  ofile<-fp(tp,paste0("topgo_",sp,".csv"))
  write.csv(res,file=ofile,row.names=FALSE,quote=FALSE)
}
```
